% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{assemble_package}
\alias{assemble_package}
\title{Takes a local package and assembles it}
\usage{
assemble_package(package_name, my_version, my_dir)
}
\arguments{
\item{package_name}{string name of the package}

\item{my_version}{string version the package should be named}

\item{my_dir}{directory path to the package}
}
\value{
No return value
}
\description{
Takes a local package and assembles it
}
\section{Side effects}{

\itemize{
  \item Modifies the version listed in the DESCRIPTION file  
  \item Deletes the tar file from any older packages
  \item Runs roxygen on the package
  \item Builds the package
  \item Tries to load the package.  If it doesn't work, the DESCRIPTION file
  version name is set back to what is was.
}
}

