% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{configure_readme}
\alias{configure_readme}
\title{Configure readme file}
\usage{
configure_readme(output_dir, file_prefix)
}
\arguments{
\item{output_dir}{path in which to save readme file}

\item{file_prefix}{value to use as prefix to _readme.txt file}
}
\description{
Method to set the README_PATH and remove existing readme file
}
