% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{detach_all_but_basic_packages}
\alias{detach_all_but_basic_packages}
\title{Detaches all packages}
\usage{
detach_all_but_basic_packages()
}
\value{
No return value.
}
\description{
Detaches all packages
}
\section{Side effects}{

\itemize{
  \item Detaches all packages
}
}

