% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{detach_package}
\alias{detach_package}
\title{Unloads packages}
\usage{
detach_package(package_name)
}
\arguments{
\item{package_name}{name of the package to detach as string}
}
\value{
No return value.
}
\description{
Unloads packages
}
\section{Side effects}{

\itemize{
  \item Unloads the named package  
}
}

