% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_along_path.R
\name{find_file_along_path}
\alias{find_file_along_path}
\title{Goes up path component and looks for the file_name in each directory.}
\usage{
find_file_along_path(my_path, file_name)
}
\arguments{
\item{my_path}{character string path to search along}

\item{file_name}{character string to search for, will match partial file names and return first match if multiple are found}
}
\value{
Path to the file if it's along the path provided
}
\description{
Goes up path component and looks for the file_name in each directory.
}
