% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_along_path.R
\name{find_folder_along_path}
\alias{find_folder_along_path}
\title{Goes up path component and looks for the folder_name in each directory.}
\usage{
find_folder_along_path(my_path, folder_name)
}
\arguments{
\item{my_path}{character string path to search along}

\item{folder_name}{character string to search for}
}
\value{
Path to the folder if it's along the path provided
}
\description{
Goes up path component and looks for the folder_name in each directory.
}
