% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{function_to_function}
\alias{function_to_function}
\title{Runs from start of fun1 and sets up args for fun2 when it gets to it.}
\usage{
function_to_function(fun_name_1, fun_name_2, to_env = globalenv())
}
\arguments{
\item{fun_name_1}{Name of the outmost function as a string.}

\item{fun_name_2}{Name of the inner function as a string.}
}
\value{
No return value.
}
\description{
Runs function_1 from the beginning of \code{fun_name_1} to the start of 
 \code{fun_name_2} and then runs \code{\link{function_var_to_env}} on that second 
 function
}
\section{Side effects}{

\itemize{
  \item runs the first line of function_1 up to the first useage of function_2 in \code{to_env}
  \item runs \code{\link{function_var_to_env}} on functnio 2 in  \code{to_env}
}
}

\section{Limitations}{

\itemize{
  \item Can't use the name to fun_name_2 in comments or anything prior to it's usage
  \item Can only go into the first usage of fun_name_2
  \item Won't work if fun_name_2 is inside of any other data structure, even if/then statements.
}
}

\section{See also}{

\itemize{
  \item \code{\link{function_var_to_env}}
}
}

