% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get_loaded_package_version}
\alias{get_loaded_package_version}
\title{Tells which version of the named package is loaded.}
\usage{
get_loaded_package_version(package_name)
}
\arguments{
\item{package_name}{string name of the package}
}
\value{
T/F on if the named package is in sessionInfo()$otherPkgs
}
\description{
Tells which version of the named package is loaded.
}
