% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{get_package_version_listed_in_description}
\alias{get_package_version_listed_in_description}
\title{Gets the package version listed in the package description file}
\usage{
get_package_version_listed_in_description(my_dir)
}
\arguments{
\item{my_dir}{directory path to the library package}
}
\value{
Version of the package found in the DESCRIPTION file
}
