% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_script_path.R
\name{get_script_path}
\alias{get_script_path}
\title{Gets the path of the script that runs it}
\usage{
get_script_path()
}
\value{
Returns the path to the script that ran it
}
\description{
Function returns the path to the script that ran it.  Can be used to put the 
script outputs right next to the script.  If run inside a function it will 
return the script that ran that function. Helps to run at the top of a sourced
function: \code{sample_script_path = get_script_path()}
And then use that script name in the function default:
\code{example_function = function(script_path = sample_script_path){script_path}}

Designed to work if run as Rscript, sourced via RStudio or console, or run 
via RStudio. Found most of the inspiration for this here: 
https://stackoverflow.com/questions/1815606/rscript-determine-path-of-the-executing-script
}
\section{Limitations}{

\itemize{
  \item Doesn't work through the console (no script there) 
  \item Works through a sourced file, but if sourced from a source script it 
  just returns the first source path.
}
}

