% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\docType{package}
\name{housekeeping-package}
\alias{housekeeping}
\alias{housekeeping-package}
\title{housekeeping: Basic Debugging and Package Building Tasks}
\description{
The housekeeping package provides tools for debugging, script path finding and package maintanance.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/DanteBortone/housekeeping/issues}
}

}
\author{
\strong{Maintainer}: Dante Bortone \email{dantebortone@gmail.com}

}
\keyword{internal}
