% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{list_missing_columns}
\alias{list_missing_columns}
\title{Prints any column names not yet defined in data}
\usage{
list_missing_columns(dat, col_names)
}
\arguments{
\item{dat}{data to look for column names within}

\item{col_names}{vector representing full set of column names expected}
}
\value{
vector of missing column names ( also printed to console )
}
\description{
Method to print out any col_names not yet defined in the data passed in. Intended to help with preparing valid data.
}
