% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{matches_loaded_version}
\alias{matches_loaded_version}
\title{Tells if my_version matches the version returned from \code{get_loaded_package_version}}
\usage{
matches_loaded_version(package_name, my_version)
}
\arguments{
\item{package_name}{string name of the package to check}

\item{my_version}{version to test if it matches the one loaded in the session}
}
\value{
T/F on if the named package is in sessionInfo()
}
\description{
Tells if my_version matches the version returned from \code{get_loaded_package_version}
}
