% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{move_to_front}
\alias{move_to_front}
\title{Move item to front of vector}
\usage{
move_to_front(my_vector, items_to_move)
}
\arguments{
\item{my_vector}{A vector}

\item{items_to_move}{Item or vector of items to move ( not indices )}
}
\value{
Returns reordered vector
}
\description{
Move item to front of vector
}
