% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{move_to_position}
\alias{move_to_position}
\title{Move item to position of vector}
\usage{
move_to_position(a_vector, items_to_move, my_position)
}
\arguments{
\item{items_to_move}{Item or vector of items to move ( not indices )}

\item{my_position}{Integer/Index of the postion the items should be moved to move the items to}

\item{my_vector}{A vector}
}
\value{
Returns reordered vector
}
\description{
Move item to position of vector
}
