% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{package_is_loaded}
\alias{package_is_loaded}
\title{Tells if package_name is currently loaded}
\usage{
package_is_loaded(package_name)
}
\arguments{
\item{package_name}{string name of the package}
}
\value{
T/F on if the named packate is in sessionInfo()$otherPkgs
}
\description{
Tells if package_name is currently loaded
}
