% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{remove_package_from_all_libraries}
\alias{remove_package_from_all_libraries}
\title{Searches all libraries in .libPaths() and deletes the named package in all of them}
\usage{
remove_package_from_all_libraries(package_name)
}
\arguments{
\item{package_name}{name of the package to detach as string}
}
\value{
No return value.
}
\section{Side effects}{

\itemize{
  \item deletes the named package in all libraries
}
}

