% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{report_df_values}
\alias{report_df_values}
\title{Turns all df or dt columns into factors and reports the counts for each.}
\usage{
report_df_values(my_df, report_columns = NULL, maxsum = 10)
}
\arguments{
\item{my_df}{data.frame or data.table from which to report values}

\item{report_columns}{Vector of columns names to check. Will default to the names in my_df.}

\item{maxsum}{What is the max number of factors to report before goes to summary}
}
\value{
No return value. Outputs message reporting the factors and counts of each column.
}
\description{
Turns all df or dt columns into factors and reports the counts for each.
}
