% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpful_functions.R
\name{use_nextflow_wd}
\alias{use_nextflow_wd}
\title{Tries to set the wd to the wd found in the Nextflow log.}
\usage{
use_nextflow_wd(script_path, script_name, set_file_pane = F)
}
\arguments{
\item{script_path}{path to being traversing down to look for a log folder}

\item{script_name}{name of script to look for in the Nextflow log}

\item{set_file_pane}{boolean indicating whether you'd like to have rstudio's file borwser pane directed to the wd}
}
\value{
vector of missing column names ( also printed to console )
}
\description{
Looks along the script path to find the Nextflow 'log' folder and looks for the script name for the last Nextflow wd
}
